Separate methods into sorting, diagnostic, canon, and async source files
