+++ src/coreclr/tools/Common/TypeSystem/Common/MethodDelegator.cs
@@ -53,7 +53,5 @@ namespace Internal.TypeSystem
 
         // For this method, delegating to the wrapped MethodDesc would likely be the wrong thing.
         public abstract override MethodDesc InstantiateSignature(Instantiation typeInstantiation, Instantiation methodInstantiation);
-
-        public abstract override AsyncMethodData AsyncMethodData { get; }
     }
 }
