+++ src/coreclr/tools/Common/TypeSystem/Common/MethodForInstantiatedType.cs
@@ -14,9 +14,6 @@ namespace Internal.TypeSystem
 
         private MethodSignature _signature;
 
-        private MethodDesc _asyncOtherVariant;
-        private AsyncMethodData _asyncMethodData;
-
         internal MethodForInstantiatedType(MethodDesc typicalMethodDef, InstantiatedType instantiatedType)
         {
             Debug.Assert(typicalMethodDef.GetTypicalMethodDefinition() == typicalMethodDef);
@@ -69,44 +66,6 @@ namespace Internal.TypeSystem
             MethodSignature template = _typicalMethodDef.Signature;
             _signature = InstantiateSignature(template);
 
-        }
-
-        public override AsyncMethodData AsyncMethodData
-        {
-            get
-            {
-                if (!_asyncMethodData.Equals(default(AsyncMethodData)))
-                    return _asyncMethodData;
-
-                if (IsAsync)
-                {
-                    // The signature should already have been updated to reflect the AsyncCallConv
-                    // No need to convert to AsyncCallConv signature
-                    Debug.Assert(!Signature.ReturnsTaskOrValueTask() && Signature.IsAsyncCallConv);
-                    _asyncMethodData = new AsyncMethodData { Kind = AsyncMethodKind.AsyncVariantImpl, Signature = Signature };
-                }
-                else if (Signature.ReturnsTaskOrValueTask())
-                {
-                    _asyncMethodData = new AsyncMethodData { Kind = AsyncMethodKind.TaskReturning, Signature = Signature };
-                }
-                else
-                {
-                    _asyncMethodData = new AsyncMethodData { Kind = AsyncMethodKind.NotAsync, Signature = Signature };
-                }
-
-                return _asyncMethodData;
-            }
-        }
-
-        public override MethodDesc GetAsyncOtherVariant()
-        {
-            if (_asyncOtherVariant is null)
-            {
-                MethodDesc otherVariant = IsAsync ? new TaskReturningAsyncThunk(this, InstantiateSignature(_typicalMethodDef.GetAsyncOtherVariant().Signature)) : new AsyncMethodThunk(this);
-                Interlocked.CompareExchange(ref _asyncOtherVariant, otherVariant, null);
-            }
-
-            return _asyncOtherVariant;
         }
 
         public override MethodSignature Signature
