+++ src/coreclr/tools/aot/ILCompiler.TypeSystem/ILCompiler.TypeSystem.csproj
@@ -42,6 +42,9 @@
     <Compile Include="..\..\Common\TypeSystem\Canon\ArrayType.Canon.cs">
       <Link>TypeSystem\Canon\ArrayType.Canon.cs</Link>
     </Compile>
+    <Compile Include="..\..\Common\TypeSystem\Canon\AsyncMethodThunk.Canon.cs">
+      <Link>TypeSystem\Canon\AsyncMethodThunk.Canon.cs</Link>
+    </Compile>
     <Compile Include="..\..\Common\TypeSystem\Canon\ByRefType.Canon.cs">
       <Link>TypeSystem\Canon\ByRefType.Canon.cs</Link>
     </Compile>
@@ -98,6 +101,9 @@
     </Compile>
     <Compile Include="..\..\Common\TypeSystem\Canon\SignatureVariable.Canon.cs">
       <Link>TypeSystem\Canon\SignatureVariable.Canon.cs</Link>
+    </Compile>
+    <Compile Include="..\..\Common\TypeSystem\Canon\TaskReturningAsyncThunk.Canon.cs">
+      <Link>TypeSystem\Canon\TaskReturningAsyncThunk.Canon.cs</Link>
     </Compile>
     <Compile Include="..\..\Common\TypeSystem\Canon\TypeDesc.Canon.cs">
       <Link>TypeSystem\Canon\TypeDesc.Canon.cs</Link>
@@ -312,6 +318,9 @@
     <Compile Include="..\..\Common\TypeSystem\Common\MethodForInstantiatedType.cs">
       <Link>TypeSystem\Common\MethodForInstantiatedType.cs</Link>
     </Compile>
+    <Compile Include="..\..\Common\TypeSystem\Common\MethodForInstantiatedType.Async.cs">
+      <Link>TypeSystem\Common\MethodForInstantiatedType.Async.cs</Link>
+    </Compile>
     <Compile Include="..\..\Common\TypeSystem\Common\MethodForInstantiatedType.Diagnostic.cs">
       <Link>TypeSystem\Common\MethodForInstantiatedType.Diagnostic.cs</Link>
     </Compile>
@@ -378,14 +387,23 @@
     <Compile Include="..\..\Common\TypeSystem\Common\MethodDelegator.cs">
       <Link>TypeSystem\Common\MethodDelegator.cs</Link>
     </Compile>
+    <Compile Include="..\..\Common\TypeSystem\Common\MethodDelegator.Async.cs">
+      <Link>TypeSystem\Common\MethodDelegator.Async.cs</Link>
+    </Compile>
     <Compile Include="..\..\Common\TypeSystem\Common\MethodDelegator.Diagnostic.cs">
       <Link>TypeSystem\Common\MethodDelegator.Diagnostic.cs</Link>
     </Compile>
     <Compile Include="..\..\Common\TypeSystem\Common\AsyncMethodThunk.cs">
       <Link>TypeSystem\Common\AsyncMethodThunk.cs</Link>
     </Compile>
+    <Compile Include="..\..\Common\TypeSystem\Common\AsyncMethodThunk.Diagnostic.cs">
+      <Link>TypeSystem\Common\AsyncMethodThunk.Diagnostic.cs</Link>
+    </Compile>
     <Compile Include="..\..\Common\TypeSystem\Common\MethodDesc.cs">
       <Link>TypeSystem\Common\MethodDesc.cs</Link>
+    </Compile>
+    <Compile Include="..\..\Common\TypeSystem\Common\MethodDesc.Async.cs">
+      <Link>TypeSystem\Common\MethodDesc.Async.cs</Link>
     </Compile>
     <Compile Include="..\..\Common\TypeSystem\Common\MethodDesc.Diagnostic.cs">
       <Link>TypeSystem\Common\MethodDesc.Diagnostic.cs</Link>
@@ -395,6 +413,9 @@
     </Compile>
     <Compile Include="..\..\Common\TypeSystem\Common\TaskReturningAsyncThunk.cs">
       <Link>TypeSystem\Common\TaskReturningAsyncThunk.cs</Link>
+    </Compile>
+    <Compile Include="..\..\Common\TypeSystem\Common\TaskReturningAsyncThunk.Diagnostic.cs">
+      <Link>TypeSystem\Common\TaskReturningAsyncThunk.Diagnostic.cs</Link>
     </Compile>
     <Compile Include="..\..\Common\TypeSystem\Common\MetadataVirtualMethodAlgorithm.cs">
       <Link>TypeSystem\Common\StandardVirtualMethodAlgorithm.cs</Link>
@@ -485,6 +506,9 @@
     </Compile>
     <Compile Include="..\..\Common\TypeSystem\Ecma\EcmaMethod.cs">
       <Link>Ecma\EcmaMethod.cs</Link>
+    </Compile>
+    <Compile Include="..\..\Common\TypeSystem\Ecma\EcmaMethod.Async.cs">
+      <Link>Ecma\EcmaMethod.Async.cs</Link>
     </Compile>
     <Compile Include="..\..\Common\TypeSystem\Ecma\EcmaMethod.Diagnostic.cs">
       <Link>Ecma\EcmaMethod.cs</Link>
@@ -579,6 +603,9 @@
     <Compile Include="..\..\Common\TypeSystem\IL\Stubs\PInvokeTargetNativeMethod.Sorting.cs">
       <Link>IL\Stubs\PInvokeTargetNativeMethod.Sorting.cs</Link>
     </Compile>
+    <Compile Include="..\..\Common\TypeSystem\Sorting\AsyncMethodThunk.Sorting.cs">
+      <Link>TypeSystem\Sorting\AsyncMethodThunk.Sorting.cs</Link>
+    </Compile>
     <Compile Include="..\..\Common\TypeSystem\Sorting\FieldDesc.Sorting.cs">
       <Link>TypeSystem\Sorting\FieldDesc.Sorting.cs</Link>
     </Compile>
@@ -593,6 +620,9 @@
     </Compile>
     <Compile Include="..\..\Common\TypeSystem\Sorting\MethodForInstantiatedType.Sorting.cs">
       <Link>TypeSystem\Sorting\MethodForInstantiatedType.Sorting.cs</Link>
+    </Compile>
+    <Compile Include="..\..\Common\TypeSystem\Sorting\TaskReturningAsyncThunk.Sorting.cs">
+      <Link>TypeSystem\Sorting\TaskReturningAsyncThunk.Sorting.cs</Link>
     </Compile>
     <Compile Include="..\..\Common\TypeSystem\Interop\FieldDesc.Interop.cs">
       <Link>TypeSystem\Interop\FieldDesc.Interop.cs</Link>
